refine internevo_monitor
